if anfrage ob die Tabelle.txt bereits vorhanden

+++ Test-Datei.py
@@ -1,12 +1,19 @@
+from pathlib import Path
 #Hier soll die Hello world Datei ausgeführt werden.
 import os
 #tabelle wird erstellt und zeilen eingetragen
 def appendtab(mdn,name,pos,dauer):
-    f = open("Tabelle.txt", "a")
+#("Tabelle.txt", "a+") Öffnet eine .txt Datei, falls sie nicht vorhanden ist, wird diese vorher erstellt
+    f = open("Tabelle.txt", "a+")
     f.write(mdn + "\t" + name + "\t" + pos + "\t" + dauer + "\n")
     f.close
+
+#die Variable f hat den pfad zur Tabelle.txt
+f = Path("C:/Users/praktikant_software/Desktop/Davins_Git_Einstieg/Tabelle.txt")
+#ist die Tabelle bereits im pfad so wird diese gelöscht
+if f.is_file():
 #die Datei "Tabelle.txt" wird gelöscht
-os.remove("Tabelle.txt")  
+    os.remove("Tabelle.txt")  
 #in der Datei Tabelle.txt werden die folgenden zeilen eingetragen  
 appendtab("1","Git","2,3","20")
 appendtab("2","Py","0","5")
